+++ packages/article/__tests__/shared-with-style.web.js
@@ -154,6 +154,7 @@ export default () => {
         onLinkPress={() => {}}
         onRelatedArticlePress={() => {}}
         onTopicPress={() => {}}
+        onTwitterLinkPress={() => {}}
         onVideoPress={() => {}}
       />
     );