+++ packages/pull-quote/src/pull-quote-twitter-link-prop-types.js
@@ -1,0 +1,10 @@
+import PropTypes from "prop-types";
+
+export const propTypes = {
+  onTwitterLinkPress: PropTypes.func.isRequired,
+  twitter: PropTypes.string
+};
+
+export const defaultProps = {
+  twitter: ""
+};